refactor: config refactor
